+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.20.0] - 2022-05-04
+
 -   `--docker-args` has been added to `dp compile`
 
 ## [0.19.0] - 2022-04-25
@@ -199,7 +201,9 @@
 -   Draft of `dp init`, `dp create`, `dp template new`, `dp template list` and `dp dbt`
 -   Draft of `dp compile` and `dp deploy`
 
-[Unreleased]: https://github.com/getindata/data-pipelines-cli/compare/0.19.0...HEAD
+[Unreleased]: https://github.com/getindata/data-pipelines-cli/compare/0.20.0...HEAD
+
+[0.20.0]: https://github.com/getindata/data-pipelines-cli/compare/0.19.0...0.20.0
 
 [0.19.0]: https://github.com/getindata/data-pipelines-cli/compare/0.18.0...0.19.0